+++ ReleaseNotes.md
@@ -1,3 +1,4 @@
 ## Release Notes - Beta 1.1
 - Completed basic CRUD features for most of the entities. 
 - Completed edit and deletion api for subject, teacher and classRoom entities, for more specific design flow diagram, refer to ![Development Document](/src/Docs/DevelopmentDocs.md)
+- Added functionality regarding import of csv, users can import teachers, classes, subjects, rooms using csv file with desginated format.